Remove all inlines
