+++ pyvrp/cpp/ProblemData.h
@@ -4,7 +4,6 @@
 #include "Matrix.h"
 #include "Measure.h"
 
-#include <cassert>
 #include <iosfwd>
 #include <limits>
 #include <optional>
@@ -261,7 +260,7 @@ public:
      *     A simple data object containing the requested location's
      *     information.
      */
-    [[nodiscard]] inline Client const &location(size_t idx) const;
+    [[nodiscard]] Client const &location(size_t idx) const;
 
     /**
      * Returns a list of all clients in the problem instance.
@@ -316,8 +315,7 @@ public:
      * VehicleType
      *     A simple data object containing the vehicle type information.
      */
-    [[nodiscard]] inline VehicleType const &
-    vehicleType(size_t vehicleType) const;
+    [[nodiscard]] VehicleType const &vehicleType(size_t vehicleType) const;
 
     /**
      * Returns the travel distance between the first and second argument,
@@ -335,7 +333,7 @@ public:
      * int
      *     Travel distance between the given clients.
      */
-    [[nodiscard]] inline Distance dist(size_t first, size_t second) const;
+    [[nodiscard]] Distance dist(size_t first, size_t second) const;
 
     /**
      * Returns the travel duration between the first and second argument,
@@ -353,7 +351,7 @@ public:
      * int
      *     Travel duration between the given clients.
      */
-    [[nodiscard]] inline Duration duration(size_t first, size_t second) const;
+    [[nodiscard]] Duration duration(size_t first, size_t second) const;
 
     /**
      * The full travel distance matrix.
@@ -364,7 +362,7 @@ public:
      *    matrix is copied, but the resulting data cannot be modified in any
      *    way!
      */
-    [[nodiscard]] inline Matrix<Distance> const &distanceMatrix() const;
+    [[nodiscard]] Matrix<Distance> const &distanceMatrix() const;
 
     /**
      * The full travel duration matrix.
@@ -375,7 +373,7 @@ public:
      *    matrix is copied, but the resulting data cannot be modified in any
      *    way!
      */
-    [[nodiscard]] inline Matrix<Duration> const &durationMatrix() const;
+    [[nodiscard]] Matrix<Duration> const &durationMatrix() const;
 
     /**
      * Number of clients in this problem instance.
@@ -473,32 +471,6 @@ public:
                 Matrix<Distance> distMat,
                 Matrix<Duration> durMat);
 };
-
-ProblemData::Client const &ProblemData::location(size_t idx) const
-{
-    assert(idx < numLocations());
-    return idx < depots_.size() ? depots_[idx] : clients_[idx - depots_.size()];
-}
-
-ProblemData::VehicleType const &
-ProblemData::vehicleType(size_t vehicleType) const
-{
-    return vehicleTypes_[vehicleType];
-}
-
-Distance ProblemData::dist(size_t first, size_t second) const
-{
-    return dist_(first, second);
-}
-
-Duration ProblemData::duration(size_t first, size_t second) const
-{
-    return dur_(first, second);
-}
-
-Matrix<Distance> const &ProblemData::distanceMatrix() const { return dist_; }
-
-Matrix<Duration> const &ProblemData::durationMatrix() const { return dur_; }
 }  // namespace pyvrp
 
 #endif  // PYVRP_PROBLEMDATA_H
